+++ dgindexer/src/main/scala/com/stratio/governance/agent/searcher/model/EntityRow.scala
@@ -1,5 +1,5 @@
 package com.stratio.governance.agent.searcher.model
 
-abstract class EntityRow(id: Int) {
-  def getId: Int = id
+abstract class EntityRow(metadataPath: String) {
+  def getMatadataPath: String = metadataPath
 }